Clearly separate server side build from scripts
