+++ package.json
@@ -36,14 +36,14 @@
     "build:client:react": "node ./scripts/run.js ./clientBuild",
     "build:client:full": "node ./scripts/run.js ./clientBuild full",
     "build:client:noDep": "node ./scripts/run.js ./clientBuild noDep",
-    "generate-fixtures": "node ./dist/generateFixtures.js",
+    "generate-fixtures": "node ./scripts/run.js ./generateFixtures",
     "lint": "eslint src playground --ignore-path playground/components/TabOverride.js",
     "playground": "shell-exec 'npm run start' 'node ./scripts/run.js ./launchBrowser http://localhost:8080/'",
     "prepublish": "npm run build",
     "start": "node ./scripts/run.js ./devServer",
     "test:unit": "mocha --compilers js:babel-core/register ./src/**/test/*.js --recursive",
     "test:unit:watch": "mocha -w --compilers js:babel-core/register ./src/**/test/*.js --recursive",
-    "test:examples": "mocha --compilers js:babel-core/register ./dist/test/**.js --recursive",
+    "test:examples": "mocha --compilers js:babel-core/register ./test/**.js --recursive",
     "test:functional": "echo 'We don't have functional tests yet",
     "test": "npm run lint && npm run build && npm run test:unit && npm run test:examples && npm run test:functional",
     "examples": "NODE_ENV=test shell-exec 'npm run start' 'node ./scripts/run.js ./launchBrowser http://localhost:8080/examples'"
